+++ sdk/storage/azure_storage_blob/src/clients/blob_service_client.rs
@@ -33,6 +33,7 @@ impl BlobServiceClient {
     ) -> Result<Self> {
         let endpoint = Url::parse(endpoint)?;
         let mut options = options.unwrap_or_default();
+        super::apply_client_defaults(&mut options.client_options);
         apply_storage_logging_defaults(&mut options.client_options);
 
         let storage_headers_policy = Arc::new(StorageHeadersPolicy);