+++ sdk/storage/azure_storage_blob/src/models/method_options.rs
@@ -180,7 +180,7 @@ pub struct BlockBlobClientUploadOptions<'a> {
 
     /// Optional. Size to partition data into.
     /// A default value will be chosen if none is provided.
-    pub partition_size: Option<NonZero<usize>>,
+    pub partition_size: Option<NonZero<u64>>,
 
     /// Optional. The server-side timeout to apply on each individual request. This is not a timeout for the whole operation.
     /// The timeout parameter is expressed in seconds. For more information, see